Remove plain text email addresses from JSON files
For marginally better security email addresses are no longer either used
in the names of JSON files or stored inside them as plain text. Instead
only hashes are stored.

JSON files are now named using the timestamp and a very rough
normalization of the commenter's name and the first few words of the
message. This needs some more work.